requirements.txt
Django==4.2.1
fastapi==0.110.3
Pillow==9.4.0
Pillow==10.3.0
PyAutoGUI==0.9.54
pyperclip==1.8.2
pytesseract==0.3.10
Requests==2.31.0
